Fixed-up existing Web API Client (the one based on Prism MVVM) - the client now enables/disables operations correctly, applies operations and updates the UI, and also now correctly supports Undo. My changes to bring the Prism code directly into the project (for my convenient inspection) has been reversed rather than committed.
The approach to Prism Mvvm command enabling/disabling was to implement a private helper function to cause all my DelegateCommands in the ViewModel to have their CanExecuteChanged event raised. Prism Mvvm will *not* cause CanExecuteChanged events to be raised automagically.

We still have a minor UI issue - some selected items can remain selected after an operation is performed (if there is still an instance of that same value left in the list.)